artifact/01-domain-collection/crt.sh/requirements.txt
psycopg2-binary
tldextract
